Rename ledwidget.py to led_widget.py
