+++ lib/python/qtvcp/widgets/led_widget.py
